requirements.txt
elevenlabs==0.2.27
python-dotenv==1.0.0
streamlit==1.29.0
pydub==0.25.1
ffmpeg-python==0.2.0